+++ NEWS.md
@@ -5,6 +5,7 @@
 - fix `transform_affine` help to remove confusion with `transforme_random_affine` help (#116, @cregouby)
 - add message translation in french (#112, @cregouby)
 - Added support for the Fashion-MNIST dataset. (#148, @koshtiakanksha)
+- Added `qmnist_dataset()` – a dataset loader for the QMNIST dataset (#153, @DerrickUnleashed)
 - Added `caltech101_dataset()` – a dataset loader for the Caltech101 dataset. (#158, @DerrickUnleashed)
 - Added `caltech256_dataset()` – a dataset loader for the Caltech256 dataset. (#158, @DerrickUnleashed)